\name{arraySize.Codelink}
\alias{arraySize.Codelink}
\title{Determine size of Codelink array}
\description{
	Determine size of Codelink array. Is used by readHeader.Codelink().
}
\usage{
    arraySize.Codelink(file, nlines)
}
\arguments{
    \item{file}{the file to be read.}
    \item{nlines}{number of lines to skip.}
}
\value{Array size.}
\details{
}
\note{
}
\author{Diego Díez Ruiz}
\seealso{}
\examples{
}
\keyword{documentation}
\keyword{utilities}
