\name{arraySize}
\alias{arraySize}
\title{Determine size of Codelink array}
\description{
	Determine size of Codelink array. Is used by readHeader().
}
\usage{
    arraySize(file, nlines)
}
\arguments{
    \item{file}{the file to be read.}
    \item{nlines}{number of lines to skip.}
}
\value{Array size.}
\details{
}
\note{
}
\author{Diego Díez}
\seealso{}
\examples{
}
\keyword{documentation}
\keyword{utilities}
