\name{codelink.exprset}
\docType{data}
\alias{codelink.exprset}
\title{Dataset of class 'CodelinkSet'}
\description{
  Dataset from a h20kcod (Codelink Human UniSet I 20k) array containing 4
  samples and ~20000 probes.
}
\usage{data(codelink.example)}
\format{
	A \code{\link[codelink:CodelinkSet-class]{Codelink}} object containing 
20469 probes and 4 samples.
}
\author{Diego Diez}
\keyword{datasets}

