\name{dec.Codelink}
\alias{dec.Codelink}
\title{Determine decimal type of Codelink files}
\description{
	Determine decimal type of Codelink files.
}
\usage{
    dec.Codelink(file, nlines)
}
\arguments{
    \item{file}{the file to be read.}
    \item{nlines}{number of lines to skip.}
}
\value{Decimal type.}
\details{
}
\note{
}
\author{Diego Díez Ruiz}
\seealso{}
\examples{
}
\keyword{documentation}
\keyword{utilities}
