\name{imageCodelink}
\alias{imageCodelink}
\title{Image plot of Codelink arrays}
\description{
  Plot and image of a Codelink array if the layout information is found.
}
\usage{
  imageCodelink(object, array=1, what="smean",
                rows=NULL, cols=NULL, col=NULL, ...)
}
\arguments{
  \item{object}{an object of class "Codelink".}
  \item{array}{array to be used.}
  \item{what}{with data plot: bg, smean, ri, ni.}
  \item{rows}{rows subset to plot.}
  \item{cols}{cols subset to plot.}
  \item{col}{colors to use.}
  \item{...}{additional arguments passed to image.}
}
\value{
}
\details{
}
\note{
}
\author{Diego Díez Ruiz}
\seealso{
}
\examples{
#    data(codelink.example)
#    imageCodelink(codelink.example)
}
\keyword{documentation}
\keyword{utilities}
