\name{Nondocumented-objects}
\alias{Nondocumented-objects}

\alias{annotation,Codelink-method}

% generic.
\alias{meanSNR}
\alias{chipDevice}
\alias{filterSNR}
\alias{codUniverse}

\alias{writeCluster}
\alias{writeCluster,ExpressionSet-method}
\alias{writeCluster,MArrayLM-method}
\alias{writeCluster,Codelink-method}
\alias{writeCluster,CodelinkSet-method}

% matrix-methods.
\alias{filterSNR,matrix-method}

% Codelink-methods.
\alias{filterSNR,Codelink-method}
\alias{codUniverse,Codelink-method}

% CodelinkSet-methods.
\alias{meanSNR,CodelinkSet-method}
\alias{chipDevice,CodelinkSet-method}
\alias{filterSNR,CodelinkSet-method}
\alias{codUniverse,CodelinkSet-method}

% non-generic functions.
\alias{Codelink2eSet}
\alias{eSet2Codelink}
\alias{guessAnnotation}
\alias{summary,Codelink-method}
\alias{summaryFlag}
\alias{summaryFlag,Codelink-method}
\alias{summaryFlag,matrix-method}


\title{Here goes every undocumented function.}
\description{
Every function without help page is redirected here. Either because it is not 
intended to have a help page or because it is a bug. If you think it is the 
later please contact the package maintainer.

NOTE: Actually it is none of that reasons but the fact that the package isunder 
heavy development.
}

\author{Diego Diez}
\keyword{documentation}
\keyword{internal}
