\name{printHead.Codelink}
\alias{printHead.Codelink}
\title{Print briefly a Codelink object}
\description{
	Takes a Codelink object and print a summary information of the data estored. It is based on printHead() from package limma.
}
\usage{
	printHead.Codelink(x)
}
\arguments{
  \item{x}{an object of class "Codelink".}
}
\value{
}
\details{
}
\note{
}
\author{Diego Diez Ruiz}
\seealso{
}
\examples{
}
\keyword{documentation}
\keyword{utilities}
