\name{printHead}
\alias{printHead}
\title{Print briefly a Codelink object}
\description{
	Takes a Codelink object and print a summary information of the data estored. It is based on printHead() from package limma.
}
\usage{
	printHead(x)
}
\arguments{
  \item{x}{an object of class "Codelink".}
}
\author{Diego Diez}
\keyword{documentation}
\keyword{utilities}
