\name{selCV.Codelink}
\alias{selCV.Codelink}
\title{Select based on CV cutoff.}
\description{
	Takes a Codelink object and select genes based on CV cutoff.
}
\usage{
	selCV.Codelink(object, cutoff)
}
\arguments{
	\item{object}{an object of class "Codelink".}
	\item{cutoff}{cutoff normaly calculated with cutCV.Codelink()}
}
\value{
	A logical vector.
}
\details{
}
\note{
}
\author{Diego Díez Ruiz}
\seealso{
}
\examples{
}
\keyword{documentation}
\keyword{utilities}
