\name{snr.Codelink}
\alias{snr.Codelink}
\title{Calculate SNR}
\description{
	Takes a Codelink object and Calculate Signal to Noise Ratio (SNR).
}
\usage{
	snr.Codelink(object)
}
\arguments{
  \item{object}{an object of class "Codelink".}
}
\value{The same Codelink object with SNR slot.}
\details{
}
\note{
}
\author{Diego Díez Ruiz}
\seealso{
}
\examples{
}
\keyword{documentation}
\keyword{utilities}
