\name{SNR}
\alias{SNR}
\title{Calculate SNR}
\description{
	Compute SNR inside read.Codelink.
}
\usage{
	SNR(Smean, Bmedian, Bstdev)
}
\arguments{
  \item{Smean}{matrix of Smean intensities.}
  \item{Bmedian}{matrix of background median intensities.}
  \item{Bstdev}{matrix of background standard deviation.}
}
\value{
}
\details{
}
\note{
}
\author{Diego Diez}
\seealso{
}
\examples{
}
\keyword{documentation}
\keyword{utilities}
