\name{write.Codelink}
\alias{write.Codelink}
\title{Write a Codelink object to file.}
\description{
}
\usage{
	write.Codelink(object, file=NULL, dec=".")
}
\arguments{
	\item{object}{an object of class "Codelink".}
	\item{file}{filename to write object to.}
	\item{dec}{decimal character to use.}
}
\value{
}
\details{
}
\note{
}
\author{Diego Díez Ruiz}
\seealso{
}
\examples{
    data(codelink.example)
    write.Codelink(codelink.example, file="codelink.example.txt")
}
\keyword{documentation}
\keyword{utilities}
