\name{arraySize}
\alias{arraySize}
\title{Determine the size of the array}
\description{
  When loading Codelink arrays in text format (as exported from the Codelink 
  software) this function retrieves the correct size of the array. This is 
  useful beacuse those files contain and indetermined number of empty lines at 
  the end. Thus, reading the entire data matrix don't works.
}
\note{
  Not meant to be used directly.
}
\author{Diego Diez}
\keyword{documentation}
\keyword{utilities}
