\name{as.matrix.Codelink}
\alias{as.matrix.Codelink}
\title{Return a matrix of intensity values}
\description{
	Takes a Codelink object and returns a matrix with
	the intensity values available.
}
\usage{
as.matrix.Codelink(x)
}
\arguments{
  \item{x}{an object of class "Codelink".}
}
\value{A matrix with the intensity values.}
\details{
}
\note{
}
\author{Diego Diez}
\seealso{\code{\link{as.matrix}}}
\examples{
#	data(codelink.example)
#	mat <- as.matrix(codelink.example)
#	is(mat)
}
\keyword{documentation}
\keyword{utilities}
