\name{codelink.example}
\docType{instance}
\alias{codelink.example}
\title{Codelink instance codelink.example}
\description{
  Subset of a dataset for 2 samples and 2000 probes.
}
\usage{data(codelink.example)}
\format{An \code{\link[codelink:Codelink-class]{Codelink}} of 2 samples.}
}
\keyword{methods}

