\name{codelink.example}
\docType{data}
\alias{codelink.example}
\title{Dataset of class 'Codelink'}
\description{
  Dataset from a h20kcod (Codelink Human UniSet I 20k) array containing 2 
  samples and ~20000 probes.
}
\usage{data(codelink.example)}
\format{
	A \code{\link[codelink:Codelink-class]{Codelink}} object containing 20469 
	probes and 2 samples.
}
\author{Diego Diez}
\keyword{datasets}

