\name{CodelinkSet-class}
\alias{class::CodelinkSet}
\alias{CodelinkSet}
\alias{CodelinkSet-class}
\docType{class}

\title{Class CodelinkSet}
\description{This is the storage class for Codelink data}
\section{Objects from the Class}{
  Object are created after reading text codelink files with readCodelink2()}

\section{Description}{
	CodelinkSet objects are derived from ExpressionSet and therefore inherits all the methods.}

\note{More details are in the package vignette}

\author{Diego Diez}

\examples{
\dontrun{
    data(codelink.exprs)
}
}
\keyword{classes}
