\name{createWeights}
\alias{createWeights}
\title{Create weight for Codelink chips}
\description{
}
\usage{
	createWeights(object, type=NULL)
}
\arguments{
	\item{object}{an object of class "Codelink".}
	\item{type}{weight assigned to each Probe\_type.}
}
\value{
}
\details{
}
\note{
}
\author{Diego Diez}
\seealso{
}
\examples{
#    data(codelink.example)
#    w <- createWeights(codelink.example, type=list(FIDUCUAL=0.1))
}
\keyword{documentation}
\keyword{utilities}
