\name{decDetect}
\alias{decDetect}
\title{Determine decimal type of Codelink files}
\description{
	Determine decimal type of Codelink files.
}
\usage{
    decDetect(file, nlines)
}
\arguments{
    \item{file}{the file to be read.}
    \item{nlines}{number of lines to skip.}
}
\value{Decimal type.}
\details{
}
\note{
}
\author{Diego Díez Ruiz}
\seealso{}
\examples{
}
\keyword{documentation}
\keyword{utilities}
