\name{Nondocumented-objects}
\alias{Nondocumented-objects}

\alias{class::CodelinkSet}
\alias{CodelinkSet}
\alias{CodelinkSet-class}

% generic.
\alias{codCorrect}
\alias{codNormalize}
\alias{getBkg}
\alias{getInt}
\alias{getInfo}
\alias{getSNR}
\alias{probeNames}
\alias{probeTypes}

% CodelinkSet-methods.
\alias{codCorrect,CodelinkSet-method}
\alias{codNormalize,CodelinkSet-method}
\alias{getBkg,CodelinkSet-method}
\alias{getInt,CodelinkSet-method}
\alias{getInfo,CodelinkSet-method}
\alias{getSNR,CodelinkSet-method}
\alias{probeNames,CodelinkSet-method}
\alias{probeTypes,CodelinkSet-method}

% non-class methods.
\alias{readCodelink2}
\alias{Codelink2eSet}
\alias{eSet2Codelink}
\alias{guessAnnotation}


\title{Here goes every undocumented function.}
\description{
Every function without help page is redirected here. Either because it is not 
intended to have a help page or because it is a bug. If you think it is the 
later please contact the package maintainer.

NOTE: Actually it is none of that reasons but the fact that the package isunder 
heavy development.
}

\author{Diego Diez}
\keyword{documentation}
\keyword{internal}
