\name{Nondocumented-objects}
\alias{Nondocumented-objects}

\alias{class::CodelinkSet}
\alias{CodelinkSet}
\alias{CodelinkSet-class}

\alias{annotation,Codelink-method}

% generic.
%\alias{arraySize}
\alias{codCorrect}
\alias{codNormalize}
\alias{codPreprocess}
\alias{getBkg}
\alias{getInt}
\alias{getInfo}
\alias{getSNR}
\alias{getType}
\alias{probeNames}
\alias{probeTypes}
\alias{meanSNR}
\alias{codPlot}
\alias{normalize.median}
\alias{chipDevice}
\alias{filterSNR}
\alias{codUniverse}
\alias{writeCluster}

% matrix-methods.
\alias{filterSNR,matrix-method}

% MArrayLM-methods.
\alias{codPlot,MArrayLM-method}
\alias{writeCluster,MArrayLM-method}

% Codelink-methods.
\alias{codPlot,Codelink-method}
\alias{filterSNR,Codelink-method}
\alias{codUniverse,Codelink-method}
\alias{writeCluster,Codelink-method}

% CodelinkSet-methods.
\alias{codCorrect,CodelinkSet-method}
\alias{codNormalize,CodelinkSet-method}
\alias{codPreprocess,CodelinkSet-method}
\alias{getBkg,CodelinkSet-method}
\alias{getInt,CodelinkSet-method}
\alias{getInfo,CodelinkSet-method}
\alias{getSNR,CodelinkSet-method}
\alias{probeNames,CodelinkSet-method}
\alias{probeTypes,CodelinkSet-method}
\alias{meanSNR,CodelinkSet-method}
\alias{codPlot,CodelinkSet-method}
\alias{exprs,CodelinkSet-method}
\alias{chipDevice,CodelinkSet-method}
\alias{filterSNR,CodelinkSet-method}
\alias{codUniverse,CodelinkSet-method}
\alias{writeCluster,CodelinkSet-method}

% non-generic functions.
\alias{readCodelink2}
\alias{readCodelinkFiles}
\alias{Codelink2eSet}
\alias{eSet2Codelink}
\alias{guessAnnotation}
\alias{summary,Codelink-method}
\alias{summaryFlag}
\alias{summaryFlag,Codelink-method}
\alias{summaryFlag,matrix-method}


\title{Here goes every undocumented function.}
\description{
Every function without help page is redirected here. Either because it is not 
intended to have a help page or because it is a bug. If you think it is the 
later please contact the package maintainer.

NOTE: Actually it is none of that reasons but the fact that the package isunder 
heavy development.
}

\author{Diego Diez}
\keyword{documentation}
\keyword{internal}
