\name{SNR}
\alias{SNR}
\title{Calculate SNR}
\description{
	Calculate SNR inside read.Codelink.
}
\usage{
	SNR(Spot_mean, Bkgd_median, Bkgd_stdev)
}
\arguments{
  \item{Spot_mean}{matrix of Spot mean intensities.}
  \item{Bkgd_median}{matrix of Bkgd median intensities.}
  \item{Bkgd_stdev}{matrix of Bkgd stdev intensities.}
}
\value{
}
\details{
}
\note{
}
\author{Diego Díez Ruiz}
\seealso{
}
\examples{
}
\keyword{documentation}
\keyword{utilities}
