\name{writeCodelink}
\alias{writeCodelink}
\title{Write a Codelink object to file.}
\description{
}
\usage{
  writeCodelink(object, file=NULL, dec=".", flag=FALSE)
}
\arguments{
  \item{object}{an object of class "Codelink".}
  \item{file}{filename to write object to.}
  \item{dec}{decimal character to use.}
  \item{flag}{should the Codelink flags be written.}
}
\value{
}
\details{
}
\note{
}
\author{Diego Diez}
\seealso{
}
\examples{
#    data(codelink.example)
#    writeCodelink(codelink.example, file="codelink.example.txt")
}
\keyword{documentation}
\keyword{utilities}
